+++ src/cljs/cxengage_javascript_sdk/domain/rest_requests.cljs
@@ -427,6 +427,49 @@
                                          (= abandon-type "count-against-sla") (assoc-in [:body :abandon-threshold] 0))]
     (api/api-request create-sla-version-request)))
 
+(defn create-integration-request [name description active integration-type properties]
+  (let [tenant-id (state/get-active-tenant-id)
+        create-integration-request (cond-> {:method :post
+                                            :url (iu/api-url "tenants/:tenant-id/integrations"
+                                                    {:tenant-id tenant-id})
+                                            :body {:active (or active false)}}
+                                      (not (nil? name))             (assoc-in [:body :name] name)
+                                      (not (nil? description))      (assoc-in [:body :description] description)
+                                      (not (nil? integration-type)) (assoc-in [:body :type] integration-type)
+                                      (not (nil? properties))       (assoc-in [:body :properties] properties))]
+    (api/api-request create-integration-request)))
+
+(defn update-integration-request [integration-id name description active properties]
+  (let [tenant-id (state/get-active-tenant-id)
+        update-integration-request (cond-> {:method :put
+                                            :url (iu/api-url "tenants/:tenant-id/integrations/:integration-id"
+                                                    {:tenant-id tenant-id :integration-id integration-id})}
+                                      (not (nil? name))         (assoc-in [:body :name] name)
+                                      (not (nil? description))  (assoc-in [:body :description] description)
+                                      (not (nil? active))       (assoc-in [:body :active] active)
+                                      (not (nil? properties))   (assoc-in [:body :properties] properties))]
+    (api/api-request update-integration-request)))
+
+(defn create-integration-listener-request [integration-id name active properties]
+  (let [tenant-id (state/get-active-tenant-id)
+        create-integration-listener-request (cond-> {:method :post
+                                                     :url (iu/api-url "tenants/:tenant-id/integrations/:integration-id/listeners"
+                                                            {:tenant-id tenant-id :integration-id integration-id})}
+                                              (not (nil? name))       (assoc-in [:body :name] name)
+                                              (not (nil? active))     (assoc-in [:body :active] active)
+                                              (not (nil? properties)) (assoc-in [:body :properties] properties))]
+    (api/api-request create-integration-listener-request)))
+
+(defn update-integration-listener-request [integration-id listener-id name active properties]
+  (let [tenant-id (state/get-active-tenant-id)
+        update-integration-listener-request (cond-> {:method :put
+                                                     :url (iu/api-url "tenants/:tenant-id/integrations/:integration-id/listeners/:listener-id"
+                                                            {:tenant-id tenant-id :integration-id integration-id :listener-id listener-id})}
+                                              (not (nil? name))       (assoc-in [:body :name] name)
+                                              (not (nil? active))     (assoc-in [:body :active] active)
+                                              (not (nil? properties)) (assoc-in [:body :properties] properties))]
+    (api/api-request update-integration-listener-request)))
+
 (defn get-layout-request [layout-id]
   (let [tenant-id (state/get-active-tenant-id)
         get-layout-request {:method :get
@@ -1253,7 +1296,7 @@
         update-transfer-list-request (cond-> {:method :put
                                               :url (iu/api-url "tenants/:tenant-id/transfer-lists/:transfer-list-id"
                                                                {:tenant-id tenant-id 
-                                                               :transfer-list-id transfer-list-id})}
+                                                                :transfer-list-id transfer-list-id})}
                                           (not (nil? name))                             (assoc-in [:body :name] name)
                                           (not (nil? description))                      (assoc-in [:body :description] description)
                                           (not (nil? active))                           (assoc-in [:body :active] active)